chore: mother of all refactors
BREAKING CHANGE: code is untested as not all of the DB functions are initialised at launch anymore
