+++ api/store/user_predictions.go
@@ -1,10 +1,9 @@
-/* storage.go
- * Contains the logic for storing user predictions and leaderboards
+/* user_predictions.go
+ * Contains the methods for interacting with the user_predictions collection
  * Authors: Zachary Bower
- * Last modified: 29/05/2025
  */
 
-package input_processing
+package store
 
 import (
 	"context"
@@ -16,33 +15,13 @@ import (
 	"go.mongodb.org/mongo-driver/mongo/options"
 )
 
-var Client *mongo.Client
-
-// Function to init DB connection from main
-// Preconditions: Receives string containing mongodb uri
-// Postconditions: Establishes connection to db located at uri and updates the global value Client to be the DB connection
-// or returns an error if it occurs
-func Init(uri string) error {
-	client, err := mongo.Connect(context.TODO(), options.Client().ApplyURI(uri))
-	if err != nil {
-		return err
-	}
-	if err := client.Ping(context.TODO(), nil); err != nil {
-		return err
-	}
-	Client = client
-	return nil
-}
-
 // Function to store user predictions in the db
 // Preconditions: receives strings containing db name, collection name and userId, and Prediction containing the users predictions
 // Postconditions: stores or updates the user's prediction stored in the db, or returns an error if the opperations was unsuccessful
-func StoreUserPrediction(dbName string, collectionName string, userId string, userPrediction Prediction) error {
-	coll := Client.Database(dbName).Collection(collectionName)
-
+func (s *Store) StoreUserPrediction(userId string, userPrediction Prediction) error {
 	// Attempt to find an existing document
 	var result Prediction
-	err := coll.FindOne(context.TODO(), bson.M{"userid": userId, "round": userPrediction.Round}).Decode(&result)
+	err := s.Collections.Predictions.FindOne(context.TODO(), bson.M{"userid": userId, "round": userPrediction.Round}).Decode(&result)
 	notFound := err == mongo.ErrNoDocuments
 
 	if err != nil && !notFound {
@@ -59,7 +38,7 @@ func StoreUserPrediction(dbName string, collectionName string, userId string, us
 
 	// The user currently does not have predictions stored so we create a new document
 	if notFound {
-		_, err := coll.InsertOne(context.TODO(), userPrediction)
+		_, err := s.Collections.Predictions.InsertOne(context.TODO(), userPrediction)
 		if err != nil {
 			return fmt.Errorf("failed to insert new user prediction: %w", err)
 		}
@@ -67,7 +46,7 @@ func StoreUserPrediction(dbName string, collectionName string, userId string, us
 	}
 
 	// Else update the user's existing prediction
-	_, err = coll.UpdateOne(context.TODO(), filter, update)
+	_, err = s.Collections.Predictions.UpdateOne(context.TODO(), filter, update)
 	if err != nil {
 		return fmt.Errorf("failed to update existing user prediction: %w", err)
 	}
@@ -77,13 +56,11 @@ func StoreUserPrediction(dbName string, collectionName string, userId string, us
 // Function to do DB lookup and get prediction for a user
 // Preconditions: receives strings containing db name, collection name and userId
 // Postconditions: returns a user's prediction if it exists, or an error if it occurs
-func GetUserPrediction(dbName string, collectionName string, userId string, round string) (Prediction, error) {
-	coll := Client.Database(dbName).Collection(collectionName)
+func (s *Store) GetUserPrediction(userId string) (Prediction, error) {
 	opts := options.FindOne()
 
 	var result Prediction
-
-	err := coll.FindOne(context.TODO(), bson.M{"userid": userId, "round": round}, opts).Decode(&result)
+	err := s.Collections.Predictions.FindOne(context.TODO(), bson.M{"userid": userId, "round": s.Round}, opts).Decode(&result)
 	if err != nil {
 		if errors.Is(err, mongo.ErrNoDocuments) {
 			return Prediction{}, err
@@ -97,14 +74,12 @@ func GetUserPrediction(dbName string, collectionName string, userId string, roun
 // Function to do DB lookup and the predictions for all users with predictions stored for a round. Used in leaderboard calculations
 // Preconditions: Receives strings containing database name, colletionname and round
 // Postconditions: Returns slice of Predictions or an error if it occurs
-func GetAllUserPredictions(dbName string, collectionName string, round string) ([]Prediction, error) {
-	coll := Client.Database(dbName).Collection(collectionName)
-
+func (s *Store) GetAllUserPredictions() ([]Prediction, error) {
 	// Filter query to match documents where the round is the round sting input to the function
-	filter := bson.D{{Key: "round", Value: round}}
+	filter := bson.D{{Key: "round", Value: s.Round}}
 	
 	// Retrieves documents that match the filter
-	cursor, err := coll.Find(context.TODO(), filter)
+	cursor, err := s.Collections.Predictions.Find(context.TODO(), filter)
 	if err != nil {
 		if errors.Is(err, mongo.ErrNoDocuments) {
 			return nil, err
@@ -119,4 +94,37 @@ func GetAllUserPredictions(dbName string, collectionName string, round string) (
 	}
 	
 	return results, nil
+}
+
+// Helper function to get valid team names used in setting user predictions. We are going to grab the valid team names
+// from the results table as this already contains a list of names, and lets us filter by round without needing to
+// create and maintain a new collection that will require more api calls
+// Preconditions: Receives db name, collection name and round strings
+// Postconditions: Returns string slice containing valid team names for the round, or returns error if an issue occurs
+func (s *Store) GetValidTeams() ([]string, string, error) {
+	// Get results stored in our db
+	dbResults, err := s.FetchMatchResultsFromDb()
+	if err != nil {
+		return nil, "", err
+	}
+
+ 	var teamNames []string
+
+    // Type assertion to determine the concrete type and extract team names
+    switch result := dbResults.(type) {
+    case SwissResultRecord:
+        // For Swiss format, Teams is map[string]string
+        for teamName := range result.Teams {
+			teamNames = append(teamNames, teamName)
+        }
+    case EliminationResultRecord:
+        // For Elimination format, Progression is map[string]*TeamProgress
+        for teamName := range result.Teams {
+            teamNames = append(teamNames, teamName)
+        }
+    default:
+        return nil, "", fmt.Errorf("unknown result record type: %T", result)
+    }
+
+    return teamNames,dbResults.GetType(), nil
 }
